Split main script into multiple files, reorder functions
