+++ unity/Assets/Scripts/AvatarContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7bc079da0d86a1543ab42c16f4748fce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
