+++ unity/Assets/Scripts/PoseDataUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d00f7facc50886418e24aa3924fbe53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
